// src/data/types.ts
export interface ISignUpFormData {
  phoneNumber: string;
  otp: string[];
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  companyName: string;
  businessID: string;
  agreedToTerms: boolean;
  creditCardName: string;
  bank: string;
  branch: string;
  accountNumber: string;
  creditCardNumber: string;
  expirationDate: string;
  cvv: string;
}